+++ app/build.gradle
@@ -1,11 +1,11 @@
 apply plugin: 'com.android.application'
 android {
-    compileSdkVersion 26
+    compileSdkVersion 27
     buildToolsVersion '27.0.3'
     defaultConfig {
         applicationId "io.treehouses.remote"
         minSdkVersion 15
-        targetSdkVersion 26
+        targetSdkVersion 27
         versionCode 2
         versionName "2.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -16,20 +16,38 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-    dependencies {
-        compile fileTree(dir: 'libs', include: ['*.jar'])
-        androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
-            exclude group: 'com.android.support', module: 'support-annotations'
-        })
-        compile 'com.android.support:appcompat-v7:26.0.1'
-        compile 'com.android.support.constraint:constraint-layout:1.0.2'
-        compile 'com.android.support:support-v4:26.0.1'
-        testCompile 'junit:junit:4.12'
-    }
     lintOptions {
         abortOnError false
     }
 }
 
 dependencies {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    androidTestImplementation('com.android.support.test.espresso:espresso-core:3.0.1', {
+        exclude group: 'com.android.support', module: 'support-annotations'
+    })
+    implementation 'com.android.support:appcompat-v7:27.1.1'
+    implementation 'com.android.support.constraint:constraint-layout:1.1.2'
+    implementation 'com.android.support:support-v4:27.1.1'
+
+    implementation 'com.android.support:multidex:1.0.3'
+    implementation 'com.afollestad.material-dialogs:commons:0.9.6.0'
+    implementation('com.mikepenz:materialdrawer:6.0.7@aar') {
+        transitive = true
+    }
+    implementation "com.android.support:recyclerview-v7:27.1.1"
+    implementation "com.android.support:support-annotations:27.1.1"
+    implementation "com.android.support:design:27.1.1"
+    implementation 'com.mikepenz:google-material-typeface:3.0.1.1.original@aar'
+    implementation 'com.mikepenz:fontawesome-typeface:5.0.6.0@aar'
+    implementation 'com.mikepenz:octicons-typeface:3.0.0.1@aar'
+    implementation 'de.hdodenhof:circleimageview:2.2.0'
+    implementation 'com.mikepenz:itemanimators:1.0.2@aar'
+    implementation 'com.mikepenz:crossfader:1.5.1@aar'
+    implementation 'com.mikepenz:crossfadedrawerlayout:1.0.1@aar'
+    implementation 'com.google.android:flexbox:1.0.0'
+    implementation "com.android.support:cardview-v7:27.1.1"
+    implementation 'com.android.support:support-v4:27.1.1'
+    testImplementation 'junit:junit:4.12'
+    androidTestImplementation 'com.android.support.test:runner:1.0.2'
 }
