+++ app/src/main/java/io/treehouses/remote/InitialActivity.java
@@ -1,0 +1,13 @@
+package io.treehouses.remote;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class InitialActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_initial);
+    }
+}
