+++ app/src/main/java/io/treehouses/remote/MainActivity.java
@@ -25,6 +25,8 @@ import android.support.v4.app.FragmentTransaction;
 import android.view.Menu;
 import android.view.MenuItem;
 
+import io.treehouses.remote.FragmentsOld.Dashboard;
+
 public class MainActivity extends FragmentActivity {
 
     public static final String TAG = "MainActivity";
