+++ app/src/main/java/io/treehouses/remote/NetworkOld/ChPasswordDialogFragment.java
@@ -1,4 +1,4 @@
-package io.treehouses.remote;
+package io.treehouses.remote.NetworkOld;
 
 import android.app.Activity;
 import android.app.AlertDialog;
@@ -11,6 +11,8 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.EditText;
+
+import io.treehouses.remote.R;
 
 /**
  * Created by going-gone on 4/19/2018.
