+++ app/src/main/java/io/treehouses/remote/NetworkOld/HotspotDialogFragment.java
@@ -1,4 +1,4 @@
-package io.treehouses.remote;
+package io.treehouses.remote.NetworkOld;
 
 import android.app.Activity;
 import android.app.AlertDialog;
@@ -8,12 +8,12 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v4.app.DialogFragment;
-import android.text.Editable;
-import android.text.TextWatcher;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.EditText;
+
+import io.treehouses.remote.R;
 
 
 /**
