+++ app/src/main/java/io/treehouses/remote/NetworkOld/WifiDialogFragment.java
@@ -1,4 +1,4 @@
-package io.treehouses.remote;
+package io.treehouses.remote.NetworkOld;
 
 import android.app.Activity;
 import android.app.AlertDialog;
@@ -12,6 +12,8 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.EditText;
+
+import io.treehouses.remote.R;
 
 /**
  * Created by yubo on 7/19/17.
