+++ app/src/main/res/values/dimens.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <!-- Define standard dimensions to comply with Holo-style grids and rhythm. -->
+
+    <dimen name="margin_tiny">4dp</dimen>
+    <dimen name="margin_small">8dp</dimen>
+    <dimen name="margin_medium">16dp</dimen>
+    <dimen name="margin_large">32dp</dimen>
+    <dimen name="margin_huge">64dp</dimen>
+
+    <!-- Semantic definitions -->
+
+    <dimen name="horizontal_page_margin">@dimen/margin_medium</dimen>
+    <dimen name="vertical_page_margin">@dimen/margin_medium</dimen>
+
+    <dimen name="fab_margin">16dp</dimen>
+    <dimen name="menu_item_icon_size">24dp</dimen>
+    <dimen name="menu_item_size">48dp</dimen>
+    <dimen name="header_height">48dp</dimen>
+    <dimen name="user_image_size">150dp</dimen>
+    <dimen name="padding_normal">8dp</dimen>
+    <dimen name="padding_large">16dp</dimen>
+    <dimen name="text_size_very_large">24sp</dimen>
+    <dimen name="text_size_large">18sp</dimen>
+    <dimen name="text_size_mid">16sp</dimen>
+    <dimen name="text_size_small">14sp</dimen>
+    <dimen name="icon_size_mid">36dp</dimen>
+    <dimen name="profile_card">150dp</dimen>
+    <dimen name="padding_small">4dp</dimen>
+    <dimen name="image_size_home_card">50dp</dimen>
+</resources>
